test: share live journey reporting and unify runner sources
Extract a reported workbench journey wrapper for Web/Desktop live specs, move live coordinator/process-runner/runner-layout to TypeScript single sources, drop parallel .mjs/.d.mts contracts, and trim harness unit tests to unique discovery/failure isolation contracts without changing the 7/2, 1/1, or 4/2 live discovery sets.
